feat: add unit tests.
